requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
PyMySQL==1.1.0
bcrypt==4.0.1
python-socketio==5.8.0
eventlet==0.33.3
python-dotenv==1.0.0
